Add SS512 and MNT224 parameters & fix polytest

+++ src/io.h
@@ -30,7 +30,7 @@ void
 hexdump(FILE * f, const string & s);
 
 void
-write_ui(string & body, unsigned int v);
+write_ui(string & body, size_t v);
 
 void
 read_ui(const unsigned char *& buf, size_t & len, unsigned int & v);
